requirements.txt
fastapi
uvicorn
faiss-cpu
sentence-transformers
torch
numpy
scikit-learn
datasets
tqdm
requests
accelerate
transformers
